Make sure all arrays have a 32bit dtype, fix typo

+++ vmlab/processes/phenology.py
@@ -130,7 +130,7 @@ class Phenology(ParameterizedProcess):
 
         self.inflo_growth_tts = np.zeros(self.GU.shape, dtype=np.float32)
         self.inflo_pheno_tts = np.zeros(self.GU.shape, dtype=np.float32)
-        self.inflo_stage = np.zeros(self.GU.shape)
+        self.inflo_stage = np.zeros(self.GU.shape, dtype=np.float32)
 
         self.full_bloom_date = np.full(self.GU.shape, np.datetime64('NaT'), dtype='datetime64[D]')
         self.DAFB = np.zeros(self.GU.shape, dtype=np.float32)
